Add tests ensuring a subspace is visible for other members
