+++ app/integration_test/main_test.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'support/appstart.dart';
 import 'tests/auth.dart';
+import 'tests/sub_spaces.dart';
 import 'tests/super_invites.dart';
 import 'tests/updates.dart';
 
@@ -11,6 +12,7 @@ void main() {
   convenientTestMain(ActerConvenientTestSlot(), () {
     group('Auth', authTests);
     group('Updates', updateTests);
+    group('Subspace', subSpaceTests);
     group('SuperInvites', superInvitesTests);
   });
 }
